gas references for ion chambers and consistent vectorized data I/O

+++ xrdmaptools/utilities/math.py
@@ -7,7 +7,7 @@ import scipy.constants as constants
 ###################################
 
 def energy_2_wavelength(energy):
-    energy = np.asarray(energy)
+    energy = np.asarray(energy).astype(float)
     # Convert energy into keV
     if np.any(energy > 1e3):
         energy /= 1e3
